+++ DaemonSvc/process_creator.h
@@ -13,7 +13,7 @@ namespace ProcessCreator
     // param:    DWORD & created_pid                if success, return created process id
     // param:    const DWORD creation_flags         see CreateProcessAsUser's parameter "dwCreationFlags" in MSDN
     // param:    const tstring & work_dir           see CreateProcessAsUser's parameter "lpCurrentDirectory" in MSDN
-    // param:    const int show_window_flag         see ShowWindow in MSDN
+    // param:    const unsigned short show_window_flag          see ShowWindow in MSDN
     // return:   HANDLE                             handle of created process
     // remarks:  implemented by Windows API "CreateProcessAsUser"
     //           @@@@@do not support Windows 2000: CreateProcessAsUser will fail with error code 1314 on Windows 2000, I don't know why
@@ -23,7 +23,7 @@ namespace ProcessCreator
         DWORD& created_pid,
         const DWORD creation_flags = CREATE_NEW_CONSOLE,
         const tstring& work_dir = TSTR(""),
-        const int show_window_flag = SW_SHOWNORMAL);
+        const unsigned short show_window_flag = SW_SHOWNORMAL);
 
     // see its overload function comment
     // param:    const DWORD pid                    source process id
@@ -32,7 +32,7 @@ namespace ProcessCreator
         DWORD& created_pid,
         const DWORD creation_flags = CREATE_NEW_CONSOLE,
         const tstring& work_dir = TSTR(""),
-        const int show_window_flag = SW_SHOWNORMAL);
+        const unsigned short show_window_flag = SW_SHOWNORMAL);
 
     // create process in current context
     // implemented by Windows API "CreateProcess"
@@ -41,6 +41,6 @@ namespace ProcessCreator
         DWORD& created_pid,
         const DWORD creation_flags = CREATE_NEW_CONSOLE,
         const tstring& work_dir = TSTR(""),
-        const int show_window_flag = SW_SHOWNORMAL);
+        const unsigned short show_window_flag = SW_SHOWNORMAL);
 }
